bot/requirements.txt
python-telegram-bot
loguru
pymongo
fuzzywuzzy
python-Levenshtein
matplotlib
numpy